RavenDB-23256 Moves `TextualFieldInserter` to dedicated file for readability.
